Refine entrypoint file (#1)
Decided to remove entrypoint file. Instead of having the docker container update the copy of the repo when it runs, it just runs with the version of the repo from the point that it was built. This means that any code changes will need to be manually deployed by rebuilding the container. This simplifies some error handling I would have had to do in the entrypoint script (what if git failed to pull?) and is just now my preferred way for this.

Also:
- Fixed log level of some logs
- Improved workflow file so it runs on all branches but only tags when on main. Formatting changes.